[video] Fix large bookmark thumbnails
To honour large advancedsettings.xml imageres values (also fix when screen size < 720).

The max thumb size is still limited because the renderer sets the scissors to the size of the
screen used by Kodi (global value from graphic context). A 4k played on 1080p screen can only be
captured in 1080p at the most.
Removing that coupling in the future would be a good idea.

Optimization: a 4k thumbnail for a 480p source is a waste of resources, with upscale processing,
storage at high resolution and later downscale for display.
Instead capture at source resolution when source dimensions < screen dimensions