Update default RSS feeds
feeds.kodi.tv is/was handled by Google Feedburner, a service that is now largely obsolete and it is unclear how long it will still be available. Switch over to our normal webpage. Since it is fronted by Cloudflare, the traffic should not be a problem.

In the course of our website revamp, the addons feeds became unavailable. Remove them since they would be largely clutter anyway.

Also, lengthen the default update interval (there aren't so many updates anyway).