[Android] Fix wake-up when sleeping with active hdmi
Improved logic for wake-up handling. If hdmi was unplugged after switching screen off, the wake-up routine is executed after hdmi is plugged in. If hdmi remained active during sleep, the wake-up routine is executed directly after receiving screen-on event (do not wait for hdmi event, which never comes in that case).